Moved thin-backend/react to thin-backend-react
